answer query reorganization
