--- mod/metrics/lib/card/fixed_company_all_answer_query.rb
@@ -1,5 +1,0 @@
-class Card
-  class FixedCompanyAllAnswerQuery < AllAnswerQuery
-    include FixedCompany
-  end
-end
