--- mod/metrics/lib/card/fixed_company_missing_answer_query.rb
@@ -1,6 +1,0 @@
-class Card
-  # find for a fixed company all metrics without metric answers
-  class FixedCompanyMissingAnswerQuery < MissingAnswerQuery
-    include FixedCompany
-  end
-end
