--- mod/metrics/lib/card/fixed_metric_all_answer_query.rb
@@ -1,5 +1,0 @@
-class Card
-  class FixedMetricAllAnswerQuery < AllAnswerQuery
-    include FixedMetric
-  end
-end
