--- mod/metrics/lib/card/fixed_metric_missing_answer_query.rb
@@ -1,6 +1,0 @@
-class Card
-  # find for a fixed metric all companies without metric answers
-  class FixedMetricMissingAnswerQuery < MissingAnswerQuery
-    include FixedMetric
-  end
-end
