+++ mod/metrics/lib/card/missing_answer_query/fixed_company.rb
@@ -1,0 +1,8 @@
+class Card
+  # find for a fixed company all metrics without metric answers
+  module MissingAnswerQuery
+    class FixedCompany < AllAnswerQuery::FixedCompany
+      include MissingAnswerQuery::Shared
+    end
+  end
+end
