+++ mod/metrics/lib/card/missing_answer_query/fixed_metric.rb
@@ -1,0 +1,8 @@
+class Card
+  # find for a fixed company all metrics without metric answers
+  module MissingAnswerQuery
+    class FixedMetric < AllAnswerQuery::FixedMetric
+      include MissingAnswerQuery::Shared
+    end
+  end
+end
