+++ mod/researched_metrics/set/type_plus_right/wikirate_company/metric_answer.rb
@@ -6,7 +6,7 @@ def virtual?
 end
 
 def query_class
-  FixedCompanyAnswerQuery
+  AnswerQuery::FixedCompany
 end
 
 def filter_card_fieldcode
